Created SplashActivity
Created SplashActivity and added the bottom sheets dialog for the user to select whether to sign up or login in or continue as guest. Created new activity for Login flow and Create Account flow. We will now handle both flows from a singular activity rather than splitting up the flows. Also removed the intent filters from StoreLocatorActivity since we were not using them and they were causing for the activity to be exported. Created new abstract activity that locks the orientation to Portrait so that we no longer have to specify it in the manifest file.